added user repository, fixed cashapi impl, fixed games, fixed reward

+++ bukkit/src/main/java/app/miyuki/miyukievents/bukkit/hook/cash/CashAPI.java
@@ -1,7 +1,9 @@
 package app.miyuki.miyukievents.bukkit.hook.cash;
 
+import java.math.BigDecimal;
+
 public interface CashAPI {
 
-    void deposit(String playerName, Double amount);
+    void deposit(String playerName, BigDecimal amount);
 
 }
