+++ bukkit/src/main/java/app/miyuki/miyukievents/bukkit/hook/cash/impl/PlayerPoints.java
@@ -5,12 +5,14 @@ import lombok.val;
 import org.black_ixx.playerpoints.PlayerPointsAPI;
 import org.bukkit.Bukkit;
 
+import java.math.BigDecimal;
+
 public class PlayerPoints implements CashAPI {
 
     private final PlayerPointsAPI cashAPI = org.black_ixx.playerpoints.PlayerPoints.getInstance().getAPI();
 
     @Override
-    public void deposit(String playerName, Double amount) {
+    public void deposit(String playerName, BigDecimal amount) {
         val uuid = Bukkit.getOfflinePlayer(playerName).getUniqueId();
 
         cashAPI.give(uuid, amount.intValue());
